Added IPEDS data files
Created a Data folder to store our datasets. The Data folder includes the downloaded IPEDS data CSV files; some Labels files for categorical variables with labels; the institution list .uid file so we can keep track of which specific institutions we're analyzing; variable list files for each of the CSV downloads so we can track the specific variables we selected for each file.