Added non-IPEDS data from WICHE and ACS
Estimates of the total number of high school graduates by state and the percent of adults with a Bachelor's or higher by state. There was no ACS 1-year estimate available for 2020...ultimately I think we may just want to use the first year for analysis purposes anyway so that might be fine.